fix - Вся логика работы с вопросами из state-data перенесена в модель game
