+++ js/game/game-model.js
@@ -1,6 +1,35 @@
 import {questions} from '../data/state-data';
-import {initialState, nextQuestionState, setNotes} from '../data/state-data';
+import {initialState} from '../data/state-data';
 import {getTimer} from '../data/game-data';
+
+
+const getQuestion = (num) => questions[num];
+
+const nextQuestionState = (state) => {
+  const currentQuestion = state.question;
+
+  let nextQuestion = currentQuestion;
+
+  if (getQuestion(currentQuestion + 1)) {
+
+    nextQuestion = currentQuestion + 1;
+
+  }
+
+  const nextState = Object.assign({}, state);
+  nextState.question = nextQuestion;
+
+  return nextState;
+};
+
+const setNotes = (state, notes) => {
+  if (notes < 0) {
+    throw new RangeError(`Can't set negative lives`);
+  }
+  state = Object.assign({}, state);
+  state.notesLeft = notes;
+  return state;
+};
 
 export default class GameModel {
   constructor(state = initialState) {
@@ -30,5 +59,9 @@ export default class GameModel {
 
   pushAnswer(answer) {
     this.state.answers.push(answer);
+  }
+
+  nextQuestionAvailable() {
+    return !!getQuestion(this.state.question + 1);
   }
 }
